+++ djangocms_helper/__init__.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
-"""djangocms-helper - Helpers for django CMS plugin development"""
 from __future__ import absolute_import, print_function, unicode_literals
 
-__version__ = '0.9.4.post2'
+__version__ = '0.9.5.post1'
 __author__ = 'Iacopo Spalletti <i.spalletti@nephila.it>'
 __all__ = ['runner']