+++ djangocms_helper/main.py
@@ -7,6 +7,7 @@ import subprocess
 import sys
 import warnings
 
+from django.utils import autoreload
 from django.utils.encoding import force_text
 from django.utils.six import text_type
 from docopt import DocoptExit, docopt
@@ -31,7 +32,7 @@ Usage:
     djangocms-helper <application> makemigrations [--extra-settings=</path/to/settings.py>] [--cms] [--merge] [--empty] [--dry-run] [--boilerplate] [<extra-applications>...]
     djangocms-helper <application> pyflakes [--extra-settings=</path/to/settings.py>] [--cms] [--boilerplate]
     djangocms-helper <application> authors [--extra-settings=</path/to/settings.py>] [--cms] [--boilerplate]
-    djangocms-helper <application> server [--port=<port>] [--bind=<bind>] [--extra-settings=</path/to/settings.py>] [--cms] [--boilerplate] [--migrate] [--no-migrate] [--persistent[=path]]
+    djangocms-helper <application> server [--port=<port>] [--bind=<bind>] [--extra-settings=</path/to/settings.py>] [--cms] [--boilerplate] [--migrate] [--no-migrate] [--persistent[=path]] [--verbose=<level>]
     djangocms-helper <application> setup [--extra-settings=</path/to/settings.py>] [--cms] [--boilerplate]
     djangocms-helper <application> <command> [options] [--extra-settings=</path/to/settings.py>] [--cms] [--boilerplate] [--migrate] [--no-migrate]
 
@@ -235,8 +236,8 @@ def static_analisys(application):
         print('Static analisys available only if django CMS is installed')
 
 
-def server(bind='127.0.0.1', port=8000, migrate_cmd=False):  # pragma: no cover
-    from django.utils import autoreload
+def server(bind='127.0.0.1', port=8000, migrate_cmd=False, verbose=1):  # pragma: no cover
+    from django.contrib.staticfiles.management.commands import runserver
 
     if os.environ.get('RUN_MAIN') != 'true':
         _create_db(migrate_cmd)
@@ -248,7 +249,6 @@ def server(bind='127.0.0.1', port=8000, migrate_cmd=False):  # pragma: no cover
             print('A admin user (username: %s, password: admin) '
                   'has been created.' % usr.get_username())
             print('')
-    from django.contrib.staticfiles.management.commands import runserver
     rs = runserver.Command()
     try:
         from django.core.management.base import OutputWrapper
@@ -261,10 +261,17 @@ def server(bind='127.0.0.1', port=8000, migrate_cmd=False):  # pragma: no cover
     rs._raw_ipv6 = False
     rs.addr = bind
     rs.port = port
+    try:
+        from channels.log import setup_logger
+        rs.logger = setup_logger('django.channels', 1)
+    except ImportError:
+        pass
     autoreload.main(rs.inner_run, (), {
         'addrport': '%s:%s' % (bind, port),
         'insecure_serving': True,
-        'use_threading': True
+        'use_threading': True,
+        'verbosity': verbose,
+        'use_reloader': True,
     })
 
 
@@ -340,7 +347,10 @@ def core(args, application):
                                             args['--runner-options'], args.get('--verbose', 1))
                         sys.exit(num_failures)
                 elif args['server']:
-                    server(args['--bind'], args['--port'], args.get('--migrate', True))
+                    server(
+                        args['--bind'], args['--port'], args.get('--migrate', True),
+                        args.get('--verbose', 1)
+                    )
                 elif args['cms_check']:
                     cms_check(args.get('--migrate', True))
                 elif args['compilemessages']: