+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.9.6 (unreleased)
+++++++++++++++++++
+
+* Added support for channels runserver.
+* Added verbosity level to server command.
+
 0.9.5 (2016-06-06)
 ++++++++++++++++++